добавил начисление баллов

+++ internal/accrual/model/orderAccrual.go
@@ -12,7 +12,7 @@ const (
 )
 
 type OrderAccrual struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual,omitempty"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual,omitempty"`
 }
